+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * Support integer values in buildUuid
   [#1375](https://github.com/bugsnag/bugsnag-android/pull/1375)
 
+* Use SystemClock.elapsedRealtime to track `app.durationInForeground`
+  [#1375](https://github.com/bugsnag/bugsnag-android/pull/1375)
+
 ## 5.12.0 (2021-08-26)
 
 * The `app.lowMemory` value always report the most recent `onTrimMemory`/`onLowMemory` status